requirements.txt
asgiref==3.4.1
Django==3.2.5
pytz==2021.1
sqlparse==0.4.1
pytest-django==4.4.0
djangorestframework==3.12.4
pytest-django==4.4.0
pytest==6.2.4